Move bulk ban test into its own class
